+++ src/Simpletools/Db/Cassandra/Doc.php
@@ -66,6 +66,7 @@ class Doc
     protected $_bubble = false;
 
 	protected $_client = null;
+    protected $_meta = null;
 
 
 //	protected $_diff = [
@@ -191,8 +192,9 @@ class Doc
 				->convertMapToJson($this->_convertMapToJson);
 
 		if($this->_ifExists) 			$this->_query->ifExists();
-		if($this->_ifNotExists) 	$this->_query->ifNotExists();
+		if($this->_ifNotExists) 	    $this->_query->ifNotExists();
 		if($this->_bubble)				$this->_query->bubble();
+        if(isset($this->_meta))		    $this->_query->setMeta($this->_meta);
 
 		return $this->_query;
 	}
@@ -233,8 +235,9 @@ class Doc
 				->convertMapToJson($this->_convertMapToJson);
 
 		if($this->_ifExists) 			$this->_query->ifExists();
-		if($this->_ifNotExists) 	$this->_query->ifNotExists();
+		if($this->_ifNotExists) 	    $this->_query->ifNotExists();
 		if($this->_bubble)				$this->_query->bubble();
+        if(isset($this->_meta))		    $this->_query->setMeta($this->_meta);
 
 		return $this->_query;
 	}
@@ -453,8 +456,9 @@ class Doc
 		}
 
 		if($this->_ifExists) 			$this->_query->ifExists();
-		if($this->_ifNotExists) 	$this->_query->ifNotExists();
+		if($this->_ifNotExists) 	    $this->_query->ifNotExists();
 		if($this->_bubble)				$this->_query->bubble();
+        if(isset($this->_meta))		    $this->_query->setMeta($this->_meta);
 
 		return $this->_query;
 	}
@@ -464,6 +468,7 @@ class Doc
 		$this->getRemoveQuery();
 		$this->_query->run();
 		$this->body(array());
+        return $this;
 	}
 
 
@@ -501,5 +506,19 @@ class Doc
 		$this->_ifExists = true;
 		return $this;
 	}
+
+  public function setMeta($meta)
+  {
+    $this->_meta = $meta;
+    return $this;
+  }
+
+  public function getMeta()
+  {
+    if($this->_query)
+      return $this->_query->getMeta();
+    else
+      return $this->_meta;
+  }
 
 }