+++ src/Simpletools/Db/Cassandra/Type/AutoIncrement.php
@@ -132,7 +132,7 @@ class AutoIncrement
         {
             (new Query(self::$_settings['table'] . '_lock'))
                     ->options([
-                        'consistency'   => Query::CONSISTENCY_ALL
+                        'consistency'   => Query::CONSISTENCY_LOCAL_QUORUM
                     ])
                     ->insertIgnore($body = [
                         'tag' => $this->_tag,
@@ -158,7 +158,7 @@ class AutoIncrement
 
                 $q = (new Query(self::$_settings['table']))
                     ->options([
-                        'consistency'   => Query::CONSISTENCY_ALL
+                        'consistency'   => Query::CONSISTENCY_LOCAL_QUORUM
                     ])
                     ->update([
                         'seq'   => new Cql('seq + 1')
@@ -169,7 +169,7 @@ class AutoIncrement
 
                 (new Query(self::$_settings['table'] . '_lock'))
                     ->options([
-                        'consistency'   => Query::CONSISTENCY_ALL
+                        'consistency'   => Query::CONSISTENCY_LOCAL_QUORUM
                     ])
                     ->delete('tag',$this->_tag)
                     ->run();